SDK update
Update SDK to version 0.0.13-alpha
